+++ integration-test/integration/chat/github_copilot_test.clj
@@ -36,8 +36,8 @@
         (match-content chat-id "system" {:type "progress" :state "running" :text "Waiting model"})
         (match-content chat-id "system" {:type "progress" :state "running" :text "Generating"})
         (match-content chat-id "assistant" {:type "text" :text "Knock "})
-        (match-content chat-id "system" {:type "usage"})
         (match-content chat-id "assistant" {:type "text" :text "knock!"})
+        (match-content chat-id "system" {:type "usage" :sessionTokens 30})
         (match-content chat-id "system" {:type "progress" :state "finished"})
         (is (match?
              {:input [{:role "user" :content [{:type "input_text" :text "Tell me a joke!"}]}]
@@ -61,9 +61,9 @@
         (match-content chat-id "user" {:type "text" :text "Who's there?\n"})
         (match-content chat-id "system" {:type "progress" :state "running" :text "Loading config"})
         (match-content chat-id "system" {:type "progress" :state "running" :text "Waiting model"})
-        (match-content chat-id "system" {:type "usage"})
         (match-content chat-id "system" {:type "progress" :state "running" :text "Generating"})
         (match-content chat-id "assistant" {:type "text" :text "Foo"})
+        (match-content chat-id "system" {:type "usage" :sessionTokens 15})
         (match-content chat-id "system" {:type "progress" :state "finished"})
         (is (match?
              {:input [{:role "user" :content [{:type "input_text" :text "Tell me a joke!"}]}
@@ -94,8 +94,8 @@
         (match-content chat-id "assistant" {:type "text" :text "Fo"})
         (match-content chat-id "assistant" {:type "text" :text "o "})
         (match-content chat-id "assistant" {:type "text" :text "bar"})
-        (match-content chat-id "system" {:type "usage"})
         (match-content chat-id "assistant" {:type "text" :text "!\n\nHa!"})
+        (match-content chat-id "system" {:type "usage" :sessionTokens 20})
         (match-content chat-id "system" {:type "progress" :state "finished"})
         (is (match?
              {:input [{:role "user" :content [{:type "input_text" :text "Tell me a joke!"}]}
@@ -137,8 +137,8 @@
         (match-content chat-id "assistant" {:type "reasonText" :id (m/pred string?) :text "y hello"})
         (match-content chat-id "assistant" {:type "reasonFinished" :id (m/pred string?) :totalTimeMs (m/pred number?)})
         (match-content chat-id "assistant" {:type "text" :text "hello "})
-        (match-content chat-id "system" {:type "usage"})
         (match-content chat-id "assistant" {:type "text" :text "there!"})
+        (match-content chat-id "system" {:type "usage" :sessionTokens 30})
         (match-content chat-id "system" {:type "progress" :state "finished"})
         (is (match?
              {:input [{:role "user" :content [{:type "input_text" :text "hello!"}]}]
@@ -169,8 +169,8 @@
         (match-content chat-id "assistant" {:type "reasonText" :id (m/pred string?) :text "ay fine"})
         (match-content chat-id "assistant" {:type "reasonFinished" :id (m/pred string?) :totalTimeMs (m/pred number?)})
         (match-content chat-id "assistant" {:type "text" :text "I'm"})
-        (match-content chat-id "system" {:type "usage"})
         (match-content chat-id "assistant" {:type "text" :text "  fine"})
+        (match-content chat-id "system" {:type "usage" :sessionTokens 30})
         (match-content chat-id "system" {:type "progress" :state "finished"})
         (is (match?
              {:input [{:role "user" :content [{:type "input_text" :text "hello!"}]}