+++ integration-test/integration/chat/google_test.clj
@@ -36,8 +36,8 @@
         (match-content chat-id "system" {:type "progress" :state "running" :text "Waiting model"})
         (match-content chat-id "system" {:type "progress" :state "running" :text "Generating"})
         (match-content chat-id "assistant" {:type "text" :text "Knoc"})
-        (match-content chat-id "system" {:type "usage"})
         (match-content chat-id "assistant" {:type "text" :text "k knock!"})
+        (match-content chat-id "system" {:type "usage" :sessionTokens 30})
         (match-content chat-id "system" {:type "progress" :state "finished"})
         (is (match?
              {:input [{:role "user" :content [{:type "input_text" :text "Tell me a joke!"}]}]
@@ -61,9 +61,9 @@
         (match-content chat-id "user" {:type "text" :text "Who's there?\n"})
         (match-content chat-id "system" {:type "progress" :state "running" :text "Loading config"})
         (match-content chat-id "system" {:type "progress" :state "running" :text "Waiting model"})
-        (match-content chat-id "system" {:type "usage"})
         (match-content chat-id "system" {:type "progress" :state "running" :text "Generating"})
         (match-content chat-id "assistant" {:type "text" :text "Foo"})
+        (match-content chat-id "system" {:type "usage" :sessionTokens 15})
         (match-content chat-id "system" {:type "progress" :state "finished"})
         (is (match?
              {:input [{:role "user" :content [{:type "input_text" :text "Tell me a joke!"}]}
@@ -93,8 +93,8 @@
         (match-content chat-id "system" {:type "progress" :state "running" :text "Generating"})
         (match-content chat-id "assistant" {:type "text" :text "Fo"})
         (match-content chat-id "assistant" {:type "text" :text "o b"})
-        (match-content chat-id "system" {:type "usage"})
         (match-content chat-id "assistant" {:type "text" :text "ar!\n\nHa!"})
+        (match-content chat-id "system" {:type "usage" :sessionTokens 20})
         (match-content chat-id "system" {:type "progress" :state "finished"})
         (is (match?
              {:input [{:role "user" :content [{:type "input_text" :text "Tell me a joke!"}]}
@@ -136,8 +136,8 @@
         (match-content chat-id "assistant" {:type "reasonText" :id (m/pred string?) :text "say hello"})
         (match-content chat-id "assistant" {:type "reasonFinished" :id (m/pred string?) :totalTimeMs (m/pred number?)})
         (match-content chat-id "assistant" {:type "text" :text "hell"})
-        (match-content chat-id "system" {:type "usage"})
         (match-content chat-id "assistant" {:type "text" :text "o there!"})
+        (match-content chat-id "system" {:type "usage" :sessionTokens 30})
         (match-content chat-id "system" {:type "progress" :state "finished"})
         (is (match?
              {:input [{:role "user" :content [{:type "input_text" :text "hello!"}]}]
@@ -168,8 +168,8 @@
         (match-content chat-id "assistant" {:type "reasonText" :id (m/pred string?) :text " say fine"})
         (match-content chat-id "assistant" {:type "reasonFinished" :id (m/pred string?) :totalTimeMs (m/pred number?)})
         (match-content chat-id "assistant" {:type "text" :text "I"})
-        (match-content chat-id "system" {:type "usage"})
         (match-content chat-id "assistant" {:type "text" :text "'m  fine"})
+        (match-content chat-id "system" {:type "usage" :sessionTokens 30})
         (match-content chat-id "system" {:type "progress" :state "finished"})
         (is (match?
              {:input [{:role "user" :content [{:type "input_text" :text "hello!"}]}